Sum of 2 Arrays- Best Solution for SDE Interviews and CP
You are given a number n1, representing the size of array a1.
You are given n1 numbers, representing elements of array a1.
You are given a number n2, representing the size of array a2.
You are given n2 numbers, representing elements of array a2.
The two arrays represent digits of two numbers.
You are required to add the numbers represented by two arrays and print the
arrays
Solution: I have attached the optimized and advanced code for this problem. 
Pro-tip: If both the arrays have same number of elements then it is very easy to sum those arrays. But, what if number of elements are different ??
Refer this solution to hack this problem.