Decimal to Any Base- Most Asked interview Ques for SDE's
We all know converting a decimal or binary to other type is a very small and less pained task. But, Irony most of the Engineers and aspiring Developers don't know a single conversion on this. 
This is best, optimized and easy way to convert base from to any other (Octa, binary, etc.) in Java. 
Do comment if you develop a more optimized way than this. Because inheriting knowledge from one another make you a Successful person in your life.